mesa: Use distro feature for checking for v6 graphics on mx6
7ulp always uses v6 graphics

+++ imx/meta-bsp/recipes-graphics/mesa/mesa_%.bbappend
@@ -4,16 +4,12 @@ PROVIDES_remove_mx7ulp = "virtual/libgles1 virtual/libgles2 virtual/libopenvg vi
 PACKAGECONFIG_remove_imxgpu2d = "gbm"
 PACKAGECONFIG_remove_mx7ulp = "egl gles gbm"
 
-do_install_append_imxgpu3d () {
-    rm -f ${D}${includedir}/GL/glx.h \
-          ${D}${includedir}/GL/glxext.h
-}
+USE_V6_GRAPHICS = "${@bb.utils.contains("DISTRO_FEATURES", "imx_v6_graphics", "yes", "no", d)}"
+USE_V6_GRAPHICS_mx7ulp = "yes"
 
-# FIXME: Dirty hack to allow use of Vivante GPU libGL binary
-do_install_append_mx7ulp () {
-    rm -f ${D}${libdir}/libGL.* \
-          ${D}${includedir}/GL/gl.h \
-          ${D}${includedir}/GL/glext.h \
-          ${D}${includedir}/GL/glx.h \
-          ${D}${includedir}/GL/glxext.h
+do_install_append_imxgpu3d () {
+    if [ "${USE_V6_GRAPHICS}" = "yes" ]; then
+       rm -f ${D}${includedir}/GL/glx.h \
+             ${D}${includedir}/GL/glxext.h
+    fi
 }
